+++ Sources/MapboxMaps/Foundation/BundleProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+internal protocol BundleProtocol: AnyObject {
+    var infoDictionary: [String: Any]? { get }
+}
+
+extension Bundle: BundleProtocol {}